+++ packages/core/client/src/schema-component/antd/form/Form.tsx
@@ -106,9 +106,18 @@ export const Form: React.FC<FormProps> & { Designer?: any } = observer(
       {
         uid: fieldSchema['x-uid'],
         async onSuccess(data) {
+          const keepValues: Record<string, any> = {};
+          Object.keys(form.fields).forEach((key) => {
+            const field = form.fields[key] as any;
+            if (field?.componentProps?.keepValueAfterSubmit) {
+              keepValues[key] = field.value;
+            } else {
+              field.initialValue = null;
+            }
+          });
           await form.reset();
-          form.setValues(data?.data);
-          form.setInitialValues(data?.data);
+          form.setValues(keepValues);
+          form.setInitialValues(keepValues);
         },
       },
       props,